+++ feature/spacex/src/main/java/com/alxnophis/jetpack/spacex/ui/viewmodel/LaunchesViewModel.kt
@@ -4,10 +4,10 @@ import androidx.lifecycle.viewModelScope
 import arrow.core.Either
 import arrow.optics.copy
 import com.alxnophis.jetpack.core.base.constants.EMPTY
-import com.alxnophis.jetpack.core.base.formatter.BaseDateFormatter
 import com.alxnophis.jetpack.core.base.provider.BaseRandomProvider
-import com.alxnophis.jetpack.core.ui.viewmodel.BaseViewModel
+import com.alxnophis.jetpack.core.ui.formatter.BaseDateFormatter
 import com.alxnophis.jetpack.core.ui.model.ErrorMessage
+import com.alxnophis.jetpack.core.ui.viewmodel.BaseViewModel
 import com.alxnophis.jetpack.spacex.R
 import com.alxnophis.jetpack.spacex.data.model.LaunchesError
 import com.alxnophis.jetpack.spacex.data.model.PastLaunchDataModel
