+++ shared/core/src/main/java/com/alxnophis/jetpack/core/di/CoreModule.kt
@@ -1,7 +1,7 @@
 package com.alxnophis.jetpack.core.di
 
-import com.alxnophis.jetpack.core.base.formatter.BaseDateFormatter
 import com.alxnophis.jetpack.core.base.provider.BaseRandomProvider
+import com.alxnophis.jetpack.core.ui.formatter.BaseDateFormatter
 import org.koin.core.module.Module
 import org.koin.dsl.module
